refactor!(serve): remove `createApp` node api
remix-serve should be a CLI, not a Node library
