+++ .changeset/grumpy-trees-flow.md
@@ -1,0 +1,7 @@
+---
+"@remix-run/serve": major
+---
+
+Remove undocumented `createApp` Node API
+
+`remix-serve` should a CLI, not a library
