--- packages/remix-serve/env.ts
@@ -1,3 +1,0 @@
-import "@remix-run/node/install";
-
-process.env.NODE_ENV = process.env.NODE_ENV || "production";
