+++ packages/remix-serve/rollup.config.js
@@ -1,64 +1,7 @@
-const path = require("node:path");
-const babel = require("@rollup/plugin-babel").default;
-const nodeResolve = require("@rollup/plugin-node-resolve").default;
-const copy = require("rollup-plugin-copy");
-
-const {
-  copyToPlaygrounds,
-  createBanner,
-  getCliConfig,
-  getOutputDir,
-  isBareModuleId,
-} = require("../../rollup.utils");
+const { getCliConfig } = require("../../rollup.utils");
 const { name: packageName, version } = require("./package.json");
 
 /** @returns {import("rollup").RollupOptions[]} */
 module.exports = function rollup() {
-  let sourceDir = "packages/remix-serve";
-  let outputDir = getOutputDir(packageName);
-  let outputDist = path.join(outputDir, "dist");
-
-  return [
-    {
-      external(id) {
-        return isBareModuleId(id);
-      },
-      input: [`${sourceDir}/index.ts`, `${sourceDir}/env.ts`],
-      output: {
-        banner: createBanner(packageName, version),
-        dir: outputDist,
-        format: "cjs",
-        preserveModules: true,
-        exports: "auto",
-      },
-      plugins: [
-        babel({
-          babelHelpers: "bundled",
-          exclude: /node_modules/,
-          extensions: [".ts", ".tsx"],
-        }),
-        nodeResolve({ extensions: [".ts", ".tsx"] }),
-        // Allow dynamic imports in CJS code to allow us to utilize
-        // ESM modules as part of the compiler.
-        {
-          name: "dynamic-import-polyfill",
-          renderDynamicImport() {
-            return {
-              left: "import(",
-              right: ")",
-            };
-          },
-        },
-        copy({
-          targets: [
-            { src: "LICENSE.md", dest: [outputDir, sourceDir] },
-            { src: `${sourceDir}/package.json`, dest: outputDir },
-            { src: `${sourceDir}/README.md`, dest: outputDir },
-          ],
-        }),
-        copyToPlaygrounds(),
-      ],
-    },
-    getCliConfig({ packageName, version }),
-  ];
+  return [getCliConfig({ packageName, version })];
 };
